re-designed the whole apis, services, removed some unncessary fields of models

+++ src/services/AdminServices.js
@@ -12,15 +12,13 @@ import { inputSanitizer } from "../middlewares/RequestValidateMiddleware.js";
 
 const ObjectID = mongoose.Types.ObjectId;
 
-export const adminLoginService = async (req, next) => {
+export const loginService = async (req, next) => {
   try {
     const reqBody = req.body;
     const data = await AdminModel.findOne({ email: reqBody.email }).exec();
-
     if (!data) {
       return { status: "fail", message: "No admin associated with this email" };
     }
-
     const isCorrectPassword = await data.isPasswordCorrect(reqBody.password);
     if (!isCorrectPassword) {
       return { status: "fail", message: "Wrong credentials" };
@@ -52,8 +50,8 @@ export const adminLoginService = async (req, next) => {
       ipAddress: req.ip,
     };
 
+    //Create a mew session
     const session = await SessionDetailsModel.create(sessionBody);
-
     // Set the sessionId to the AdminModel
     data.sessionId.push(session._id);
     await data.save();
@@ -76,8 +74,15 @@ export const adminLoginService = async (req, next) => {
   }
 };
 
-export const adminProfileDetailsService = async (req, next) => {
+export const getProfileService = async (req, next) => {
   try {
+    const userID = req.params.id;
+    // SuperAdmin can check any admin's profile if by sending the userId into query
+    if (req.headers.role === "SuperAdmin" && userID) {
+      const data = await AdminModel.findById(userID).select("-password");
+      return { status: "success", data: data };
+    }
+    // Get the profile by own userId
     const data = await AdminModel.findOne({ _id: req.headers.id }).select(
       "-password"
     );
@@ -90,15 +95,18 @@ export const adminProfileDetailsService = async (req, next) => {
   }
 };
 
-export const addNewAdminService = async (req, next) => {
+export const createAdminService = async (req, next) => {
   try {
     let reqBody = req.body;
+    //Check if account already exists
     const existingAdmin = await AdminModel.findOne({
       email: reqBody.email,
     }).exec();
+
     if (existingAdmin) {
       return { status: "fail", message: "Admin already exists" };
     }
+    //Create a new admin if not exists
     const data = await AdminModel.create(reqBody);
     if (!data) {
       return { status: "fail", message: "Failed to add new admin" };
@@ -112,44 +120,65 @@ export const addNewAdminService = async (req, next) => {
   }
 };
 
-export const deleteAdminService = async (req, next) => {
+export const updateAdminService = async (req, next) => {
   try {
-    let id = req.query.id;
-    const data = await AdminModel.deleteOne({ _id: id });
-    if (!data) {
-      return { status: "fail", message: "Failed to delete admin" };
+    //Using fineOne instead of findOneAndUpdate beacuse findOneAndUpdate doesn't trigger the pre "save" which is a must for hashing
+    const admin = await AdminModel.findById(req.headers.id).select("-password");
+    if (!admin) {
+      return { status: "fail", message: "No profile found" };
     }
-    return { status: "success", data: data };
+    //Assign the req.body values to admin document
+    Object.assign(admin, req.body);
+
+    const updatedAdmin = await admin.save();
+
+    return {
+      status: "success",
+      message: "Profile details updated",
+      data: updatedAdmin,
+    };
   } catch (error) {
     next(error);
   }
 };
 
-export const adminListService = async (req, next) => {
+export const deleteAdminService = async (req, next) => {
+  try {
+    const data = await AdminModel.deleteOne({ _id: req.params.id });
+    if (!data || data.deletedCount < 1) {
+      return { status: "fail", message: "Failed to delete admin" };
+    }
+    return { status: "success", message: "Admin has been deleted", data: [] };
+  } catch (error) {
+    next(error);
+  }
+};
+
+export const getAdminListService = async (req, next) => {
   try {
     const data = await AdminModel.find({ role: "Admin" });
     if (!data) {
       return { status: "fail", message: "Failed to load admin list" };
     }
-    return { status: "success", data: data };
+    return { status: "success", total: data.length, data: data };
   } catch (error) {
     next(error);
   }
 };
 
-export const userListService = async (req, next) => {
+export const getUserListService = async (req, next) => {
   try {
     const data = await UserModel.find({ role: "User" });
     if (!data) {
       return { status: "fail", message: "Failed to load user list" };
     }
-    return { status: "success", data: data };
+    return { status: "success", total: data.length, data: data };
   } catch (error) {
     next(error);
   }
 };
 
-export const reviewPostListService = async (req, next) => {
+export const getReviewPostListService = async (req, next) => {
   try {
     const data = await PostModel.find({ onReview: true, isApproved: false });
     if (!data) {
@@ -161,13 +190,45 @@ export const reviewPostListService = async (req, next) => {
   }
 };
 
-/*Admin will only show those post which he approved*/
-export const approvedPostListService = async (req, next) => {
+export const getReviewPostIdsService = async (req, next) => {
   try {
+    const data = await PostModel.aggregate([
+      {
+        $match: { onReview: true, isApproved: false },
+      },
+      { $sort: { createdAt: -1 } },
+      {
+        $project: {
+          _id: 1,
+        },
+      },
+    ]);
+    if (!data) {
+      return { status: "fail", message: "Failed to load review post list" };
+    }
+    const ids = data.map((item) => item._id.toString());
+    return { status: "success", total: ids.length, data: { id: ids } };
+  } catch (error) {
+    next(error);
+  }
+};
+
+export const getApprovedPostListService = async (req, next) => {
+  try {
+    /*SuperAdmin can view all the approved post but admin can only view those account which they approved*/
+    //For SuperAdmins
+    if (req.headers.role === "SuperAdmin") {
+      const data = await PostModel.find({
+        isApproved: true,
+        onReview: false,
+      });
+      return { status: "success", total: data.length, data: data };
+    }
+    //For Admins
     const data = await PostModel.find({
       isApproved: true,
       onReview: false,
-      "approvedBy.id": new ObjectID(req.headers.id),
+      "approvedBy.userId": new ObjectID(req.headers.id),
     });
     if (!data) {
       return { status: "fail", message: "Failed to load approve post list" };
@@ -178,12 +239,23 @@ export const approvedPostListService = async (req, next) => {
   }
 };
 
-/*Admin will only show those post which he declined*/
-export const declinedPostListService = async (req, next) => {
+export const getDeclinedPostListService = async (req, next) => {
   try {
+    /*SuperAdmin can view all the declined post but admin can only view those account which they declined*/
+    //For SuperAdmins
+    if (req.headers.role === "SuperAdmin") {
+      const data = await PostModel.find({
+        isApproved: false,
+        onReview: false,
+        isDeclined: true,
+      });
+      return { status: "success", total: data.length, data: data };
+    }
     const data = await PostModel.find({
       isDeclined: true,
-      "declinedBy.id": new ObjectID(req.headers.id),
+      isApproved: false,
+      onReview: false,
+      "declinedBy.userId": new ObjectID(req.headers.id),
     });
     if (!data) {
       return { status: "fail", message: "Failed to load declined post list" };
@@ -194,7 +266,7 @@ export const declinedPostListService = async (req, next) => {
   }
 };
 
-export const reportedPostListService = async (req, next) => {
+export const getReportedPostListService = async (req, next) => {
   try {
     const data = await PostModel.find({ reportCount: { $gte: 1 } });
     if (!data) {
@@ -208,18 +280,28 @@ export const reportedPostListService = async (req, next) => {
 
 export const withdrawReportService = async (req, next) => {
   try {
-    const postID = req.query.postId;
     const data = await PostModel.findOneAndUpdate(
-      { _id: postID },
-      { $set: { reportAdmin: false, reportedBy: "" } },
+      { _id: req.params.postId },
+      {
+        $pull: { reportedBy: { _id: req.body.reportId } },
+        $inc: { reportCount: -1 },
+      },
       {
         new: true,
       }
     );
+
     if (!data) {
       return { status: "fail", message: "Failed to withdraw report" };
     }
-    return { status: "success", total: data.length, data: data };
+
+    // Manually ensure reportCount does not go below zero
+    if (data.reportCount < 0) {
+      data.reportCount = 0;
+      await data.save();
+    }
+
+    return { status: "success", data: data };
   } catch (error) {
     next(error);
   }
@@ -232,15 +314,13 @@ export const approvePostService = async (req, next) => {
             "66a3de66e505fd46ea7c2435",
             "66a3de66e505fd46ea7c2435"
         ]
-
         or
-
         "postId": [
             "66a3de66e505fd46ea7c2435"
         ]
           */
 
-    const postId = req.body.postId;
+    const postId = req.params.postId;
     const { id, name, role } = req.headers;
     inputSanitizer(postId);
 
@@ -252,30 +332,17 @@ export const approvePostService = async (req, next) => {
           onReview: false,
           declinedBy: "",
           isApproved: true,
-          "approvedBy.id": new ObjectID(id),
+          "approvedBy.userId": new ObjectID(id),
           "approvedBy.name": name,
           "approvedBy.role": role,
         },
       }
     );
-
     if (data.modifiedCount === 0) {
       return { status: "fail", message: "Failed to approve post" };
     }
 
-    const adminResponse = await AdminModel.findOneAndUpdate(
-      { _id: id },
-      { $addToSet: { approvedPosts: { $each: postId } } }
-    );
-
-    if (!adminResponse) {
-      return {
-        status: "fail",
-        message: "Failed to update admin model with approved posts",
-      };
-    }
-
-    return { status: "success", data: data };
+    return { status: "success", data: [], message: "Posts Approved" };
   } catch (error) {
     next(error);
   }
@@ -288,9 +355,7 @@ export const declinePostService = async (req, next) => {
             "66a3de66e505fd46ea7c2435",
             "66a3de66e505fd46ea7c2435"
         ]
-
         or
-
         "postId": [
             "66a3de66e505fd46ea7c2435"
         ]
@@ -309,7 +374,7 @@ export const declinePostService = async (req, next) => {
           onReview: false,
           approvedBy: "",
           isDeclined: true,
-          "declinedBy.id": new ObjectID(id),
+          "declinedBy.userId": new ObjectID(id),
           "declinedBy.name": name,
           "declinedBy.role": role,
         },
@@ -341,47 +406,56 @@ export const declinePostService = async (req, next) => {
 export const deletePostService = async (req, next) => {
   const session = await mongoose.startSession();
   try {
-    const postID = req.query.postId;
+    const postID = req.params.postId;
+
+    // Start the transaction
     session.startTransaction();
 
+    // Delete post
     const post = await PostModel.deleteOne({ _id: postID }).session(session);
 
+    if (!post) {
+      return { status: "fail", message: "Post not found" };
+    }
+
     if (post.deletedCount !== 1) {
+      // If post delete fails, abort transaction
       await session.abortTransaction();
-      session.endSession();
       return { status: "fail", message: "Failed to delete post by admin" };
     }
 
+    // Delete post details
     const postDetails = await PostDetailsModel.deleteOne({
       postID: postID,
     }).session(session);
 
     if (postDetails.deletedCount !== 1) {
+      // If post details delete fails, abort transaction
       await session.abortTransaction();
-      session.endSession();
       return {
         status: "fail",
         message: "Failed to delete post details by admin",
       };
     }
 
-    session.commitTransaction();
-    session.endSession();
+    // If all deletes succeeded, commit the transaction
+    await session.commitTransaction();
 
     return { status: "success", message: "Post deleted by admin successfully" };
   } catch (error) {
+    // Abort the transaction on any error
     await session.abortTransaction();
-    session.endSession();
     next(error);
     return { status: "fail", message: "Something went wrong" };
   } finally {
-    session.endSession();
+    // Ensure session is ended only once in the `finally` block
+    await session.endSession();
   }
 };
 
 export const sendFeedbackService = async (req, next) => {
   try {
-    const postID = req.query.postId;
+    const postID = req.params.postId;
     const reqBody = req.body;
 
     inputSanitizer(reqBody);
@@ -410,7 +484,7 @@ export const sendFeedbackService = async (req, next) => {
   }
 };
 
-export const warnedAccountListService = async (req, next) => {
+export const getWarnedAccountListService = async (req, next) => {
   try {
     const data = await UserModel.find({ accountStatus: "Warning" });
     if (!data) {
@@ -422,7 +496,7 @@ export const warnedAccountListService = async (req, next) => {
   }
 };
 
-export const restrictedAccountListService = async (req, next) => {
+export const getRestrictedAccountListService = async (req, next) => {
   try {
     const data = await UserModel.find({ accountStatus: "Restricted" });
     if (!data) {
@@ -439,9 +513,9 @@ export const restrictedAccountListService = async (req, next) => {
 
 export const withdrawRestrictionsService = async (req, next) => {
   try {
-    let id = req.query.id;
+    const { userId } = req.params;
     const data = await UserModel.findOneAndUpdate(
-      { _id: id },
+      { _id: userId },
       { $set: { accountStatus: "Validate" } },
       { new: true }
     );
@@ -454,40 +528,11 @@ export const withdrawRestrictionsService = async (req, next) => {
   }
 };
 
-export const restrictAccountService = async (req, next) => {
-  try {
-    let id = req.query.id;
-    const data = await UserModel.findOneAndUpdate(
-      { _id: id },
-      { $set: { accountStatus: "Restricted" } },
-      { new: true }
-    );
-    if (!data) {
-      return { status: "fail", message: "Failed to restrict account" };
-    }
-
-    const adminResponse = await AdminModel.findOneAndUpdate(
-      { _id: req.headers.id },
-      { $addToSet: { restrictedAccounts: id } }
-    );
-    if (!adminResponse) {
-      return {
-        status: "fail",
-        message: "Failed to restrict account, check admin model",
-      };
-    }
-
-    return { status: "success", data: data };
-  } catch (error) {
-    next(error);
-  }
-};
-
 export const warningAccountService = async (req, next) => {
   try {
-    let id = req.query.id;
+    const { userId } = req.params;
     const data = await UserModel.findOneAndUpdate(
-      { _id: id },
+      { _id: userId },
       { $set: { accountStatus: "Warning" }, $inc: { warningCount: 1 } },
       { new: true }
     );
@@ -497,7 +542,7 @@ export const warningAccountService = async (req, next) => {
 
     const adminResponse = await AdminModel.findOneAndUpdate(
       { _id: req.headers.id },
-      { $addToSet: { warnedAccounts: id } }
+      { $addToSet: { warnedAccounts: userId } }
     );
     if (!adminResponse) {
       return {
@@ -512,7 +557,36 @@ export const warningAccountService = async (req, next) => {
   }
 };
 
-export const reviewNidListService = async (req, next) => {
+export const restrictAccountService = async (req, next) => {
+  try {
+    const { userId } = req.params;
+    const data = await UserModel.findOneAndUpdate(
+      { _id: userId },
+      { $set: { accountStatus: "Restricted" } },
+      { new: true }
+    );
+    if (!data) {
+      return { status: "fail", message: "Failed to restrict account" };
+    }
+
+    const adminResponse = await AdminModel.findOneAndUpdate(
+      { _id: req.headers.id },
+      { $addToSet: { restrictedAccounts: userId } }
+    );
+    if (!adminResponse) {
+      return {
+        status: "fail",
+        message: "Failed to restrict account, check admin model",
+      };
+    }
+
+    return { status: "success", data: data };
+  } catch (error) {
+    next(error);
+  }
+};
+
+export const getReviewNidListService = async (req, next) => {
   try {
     const data = await UserModel.find({
       nidSubmitted: true,
@@ -538,7 +612,7 @@ export const reviewNidListService = async (req, next) => {
 
 export const approveNidService = async (req, next) => {
   try {
-    let userID = req.query.userId;
+    let userID = req.params.userId;
     const data = await UserModel.findOneAndUpdate(
       {
         _id: userID,
@@ -550,7 +624,8 @@ export const approveNidService = async (req, next) => {
         $set: {
           nidVerified: true,
         },
-      }
+      },
+      { new: true }
     );
 
     if (!data) {
@@ -572,7 +647,7 @@ export const approveNidService = async (req, next) => {
 
 export const declineNidService = async (req, next) => {
   try {
-    let userID = req.query.userId;
+    let userID = req.params.userId;
     const data = await UserModel.findOneAndUpdate(
       {
         _id: userID,
@@ -603,29 +678,6 @@ export const declineNidService = async (req, next) => {
       message: "NID request Declined",
       data: data,
     };
-  } catch (error) {
-    next(error);
-  }
-};
-
-export const reviewPostListIdOnlyService = async (req, next) => {
-  try {
-    const data = await PostModel.aggregate([
-      {
-        $match: { onReview: true, isApproved: false },
-      },
-      { $sort: { createdAt: -1 } },
-      {
-        $project: {
-          _id: 1,
-        },
-      },
-    ]);
-    if (!data) {
-      return { status: "fail", message: "Failed to load review post list" };
-    }
-    const ids = data.map((item) => item._id.toString());
-    return { status: "success", total: ids.length, data: { id: ids } };
   } catch (error) {
     next(error);
   }
@@ -683,24 +735,6 @@ export const searchAdminService = async (req, next) => {
       return { status: "fail", message: "No account found" };
     }
     return { status: "success", total: response.length, data: response };
-  } catch (error) {
-    next(error);
-  }
-};
-
-export const updateAdminProfileService = async (req, next) => {
-  try {
-    const userId = req.headers.id;
-    const reqBody = req.body;
-    const response = await AdminModel.findOneAndUpdate(
-      { _id: userId },
-      { $set: reqBody },
-      { upsert: true, new: true }
-    );
-    if (!response) {
-      return { status: "fail", message: "No profile found" };
-    }
-    return { status: "success", data: response };
   } catch (error) {
     next(error);
   }
